Create AWS-Mysql .yml
I have created a AWS template for the RDS server. This is not entended for use as is and will need to be combined with  the AWS template that includes the correct network and ec2 server. Once that template is done i will need to tweak a few settings so that it is using that network but outside of that this currently works for creating an RDS database of mysql8 with the appropriate MySQL settings.