Added database to templates
The AWS-UbuntuRDS.yml template creates an EC2 
server and a stand-alone RDS server running mysql, 
while the AWS-Ubiuntu.yml template creates an EC2 
server that installs mysql when it's setup. The second 
has the advantage of being cheaper to maintain and 
run at the expense of a slight performance hit.