Added connections png and database schema for spiral 3
This commit adds a png showing all of the connections and connection ids for spiral 3. It also includes a new SQL file with the most recent schema changes talked about for spiral 3/4. I renamed the existing SQL file to wsutunnelapp_spiral2_228.sql to better reflect what it is in case someoen needs to pull the old data.